don't clean quite everything

+++ external/nvidia-graphics-drivers/clean.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 for i in $(svn stat | grep '\?' | awk '{print $2}'); do
+	[[ "$i" = "NVIDIA-Linux-x86-1.0-8178-pkg1.run" ]] && continue
 	rm -rf "$i"
 done
